requirements.txt
fastapi==0.70.0
uvicorn==0.15.0
pydantic
passlib==1.7.4
bcrypt==3.2.0
python-jose==3.3.0
PyJWT == 2.3.0
pydantic[email]
prometheus_client
python-multipart